+++ src/app/(common)/mypage/_components/ProfileSection/ProfileEditModal.tsx
@@ -4,81 +4,17 @@ import Modal from "@/components/Modal";
 import Button from "@/components/Button";
 import Input from "@/components/Input";
 import Edit from "@/images/edit.svg";
-import { useForm } from "react-hook-form";
-import { useUpdateUserInfo } from "@/app/(common)/mypage/_hooks/useUpdateUserInfo";
 import Avatar from "@/components/Avatar";
-import { useEffect, useState } from "react";
-import { useGetUserInfo } from "@/app/(common)/mypage/_hooks/useGetUserInfo";
+import { useProfileEditForm } from "@/app/(common)/mypage/_hooks/useProfileEditForm";
 
 type ProfileEditModalProps = {
   isOpen: boolean;
   onClose: () => void;
 };
 
-// 회원 정보 폼 타입
-type FormValues = {
-  companyName: string;
-  profileImg: FileList;
-};
-
 // 프로필을 수정하는 폼
 export default function ProfileEditModal({ isOpen, onClose }: ProfileEditModalProps) {
-  const { data } = useGetUserInfo();
-  const [previewUrl, setPreviewUrl] = useState(data?.image || "");
-  const {
-    register,
-    handleSubmit,
-    watch,
-    reset,
-    formState: { isValid, errors },
-  } = useForm<FormValues>({ mode: "onChange", defaultValues: { companyName: data?.companyName } });
-  const mutation = useUpdateUserInfo();
-
-  // 이미지 상태를 추적
-  const watchedProfileImg = watch("profileImg");
-
-  // 이미지 미리보기 설정
-  useEffect(() => {
-    if (watchedProfileImg && watchedProfileImg.length > 0) {
-      const file = watchedProfileImg[0];
-      // 이미지 파일를 URL로 변환
-      const objectUrl = URL.createObjectURL(file);
-      setPreviewUrl(objectUrl);
-
-      // 파일이 변경되거나 컴포넌트가 unmount될 때 URL 해제
-      return () => URL.revokeObjectURL(objectUrl);
-    } else {
-      // 파일 선택이 없으면 기존 imageUrl 사용
-      setPreviewUrl(data?.image || "");
-    }
-  }, [watchedProfileImg, data?.image]);
-
-  // 모달을 여닫을 때 상태를 초기화
-  useEffect(() => {
-    if (isOpen) {
-      reset({ companyName: data?.companyName, profileImg: undefined });
-      setPreviewUrl(data?.image || "");
-    }
-  }, [isOpen, data?.companyName, data?.image, reset]);
-
-  // 폼 제출 로직
-  const onSubmit = (data: FormValues) => {
-    const formData = new FormData();
-    formData.append("companyName", data.companyName);
-
-    if (data.profileImg && data.profileImg.length > 0) {
-      formData.append("image", data.profileImg[0]);
-    }
-    mutation.mutate(formData);
-    onClose();
-  };
-
-  const isProfileUnchanged = (!watchedProfileImg || watchedProfileImg.length === 0) && previewUrl === data?.image;
-  const isCompanyNameUnchanged = watch("companyName") === data?.companyName;
-
-  // 버튼 비활성화 조건
-  const isDisabled = !isValid || (isProfileUnchanged && isCompanyNameUnchanged);
-
+  const { register, handleSubmit, onSubmit, previewUrl, errors, isDisabled } = useProfileEditForm(isOpen, onClose);
   return (
     <Modal isOpen={isOpen} onClose={onClose} className="mx-4 flex w-full max-w-[520px] flex-col gap-6 md:mx-0">
       <div className="text-lg font-semibold">프로필 수정하기</div>
